feat(provider): Billing settings - Payer directory Candid
