+++ examples/medplum-provider/src/utils/billing.test.ts
@@ -3,31 +3,22 @@
 import type { Organization, Parameters } from '@medplum/fhirtypes';
 import { describe, expect, test } from 'vitest';
 import {
-  BILLING_ORGANIZATION_IDENTIFIER_VALUE,
   CANDID_ELIGIBILITY_PAYER_ID_SYSTEM,
   CANDID_ELIGIBILITY_SUPPORT_EXTENSION,
   CANDID_PAYER_CATEGORY_SYSTEM,
   CANDID_PAYER_UUID_SYSTEM,
   CHC_PAYER_ID_SYSTEM,
   CMS_PAYER_ID_SYSTEM,
-  EIN_SYSTEM,
-  MEDPLUM_PROVIDER_IDENTIFIER_SYSTEM,
-  NPI_SYSTEM,
   ORGANIZATION_TYPE_SYSTEM,
   PAYER_ORGANIZATION_TYPE,
-  PROVIDER_ORGANIZATION_TYPE,
   buildPayerRefreshOps,
-  buildUpdatedOrganization,
   formatPayerCategory,
   getPayerCategory,
   getPayerId,
   getPayerUuid,
   isPayerNotFoundError,
-  isValidBillingPhone,
-  isValidNpi,
   parsePayerSearchPage,
   upsertIdentifier,
-  upsertPhone,
 } from './billing';
 
 // A payer Organization as the candid-get-payers bot builds it from Candid's payers.v4 API.
@@ -46,171 +37,28 @@ function makeDirectoryPayer(overrides: Partial<Organization> = {}): Organization
 }
 
 describe('billing utils', () => {
-  describe('isValidNpi', () => {
-    test.each(['1234567893', '3564119220', '1679576722', '7778889999'])('accepts valid NPI %s', (npi) => {
-      expect(isValidNpi(npi)).toBe(true);
-    });
-
-    test.each([
-      ['1234567890', 'wrong check digit'],
-      ['7778889990', 'wrong check digit'],
-      ['123456789', 'too short'],
-      ['12345678931', 'too long'],
-      ['123456789X', 'non-digit'],
-      ['', 'empty'],
-    ])('rejects %s (%s)', (npi) => {
-      expect(isValidNpi(npi)).toBe(false);
-    });
-  });
-
-  describe('isValidBillingPhone', () => {
-    test('accepts formatted 10-digit numbers', () => {
-      expect(isValidBillingPhone('(212) 555-1234')).toBe(true);
-      expect(isValidBillingPhone('9177773344')).toBe(true);
-    });
-
-    test('rejects numbers starting with 0 or 1', () => {
-      expect(isValidBillingPhone('0125551234')).toBe(false);
-      expect(isValidBillingPhone('1234567890')).toBe(false);
-    });
-
-    test('rejects wrong lengths', () => {
-      expect(isValidBillingPhone('212555123')).toBe(false);
-      expect(isValidBillingPhone('21255512345')).toBe(false);
-      expect(isValidBillingPhone('')).toBe(false);
-    });
-  });
-
   describe('upsertIdentifier', () => {
     test('appends a new identifier', () => {
-      expect(upsertIdentifier(undefined, NPI_SYSTEM, '3564119220')).toEqual([
-        { system: NPI_SYSTEM, value: '3564119220' },
+      expect(upsertIdentifier(undefined, CHC_PAYER_ID_SYSTEM, '60054')).toEqual([
+        { system: CHC_PAYER_ID_SYSTEM, value: '60054' },
       ]);
     });
 
     test('replaces the value for the same system, preserving other systems and extra props', () => {
       const identifiers = [
         { system: 'https://example.com/other', value: 'keep-me' },
-        { system: NPI_SYSTEM, value: '1234567893', use: 'official' as const },
+        { system: CHC_PAYER_ID_SYSTEM, value: '60054', use: 'official' as const },
       ];
-      const result = upsertIdentifier(identifiers, NPI_SYSTEM, '3564119220');
+      const result = upsertIdentifier(identifiers, CHC_PAYER_ID_SYSTEM, '62308');
       expect(result).toEqual([
         { system: 'https://example.com/other', value: 'keep-me' },
-        { system: NPI_SYSTEM, value: '3564119220', use: 'official' },
+        { system: CHC_PAYER_ID_SYSTEM, value: '62308', use: 'official' },
       ]);
     });
 
     test('removes the identifier on empty value', () => {
-      const identifiers = [{ system: NPI_SYSTEM, value: '1234567893' }];
-      expect(upsertIdentifier(identifiers, NPI_SYSTEM, '  ')).toBeUndefined();
-    });
-  });
-
-  describe('upsertPhone', () => {
-    test('replaces the phone entry, preserving email entries', () => {
-      const telecom = [
-        { system: 'email' as const, value: 'billing@example.com' },
-        { system: 'phone' as const, value: '0000000000', use: 'work' as const },
-      ];
-      expect(upsertPhone(telecom, '2125551234')).toEqual([
-        { system: 'email', value: 'billing@example.com' },
-        { system: 'phone', value: '2125551234', use: 'work' },
-      ]);
-    });
-
-    test('appends when no phone exists and removes on empty', () => {
-      expect(upsertPhone(undefined, '2125551234')).toEqual([{ system: 'phone', value: '2125551234' }]);
-      expect(upsertPhone([{ system: 'phone', value: '2125551234' }], '')).toBeUndefined();
-    });
-  });
-
-  describe('buildUpdatedOrganization', () => {
-    test('builds a new billing organization with prov type, identifiers, and digits-only EIN', () => {
-      const result = buildUpdatedOrganization(
-        { resourceType: 'Organization' },
-        {
-          name: ' Test Medical Practice LLC ',
-          npi: '3564119220',
-          ein: '12-3456789',
-          phone: '6175550142',
-          address: { line: ['456 Medical Center Drive'], city: 'Boston', state: 'MA', postalCode: '02101' },
-        }
-      );
-
-      expect(result.name).toBe('Test Medical Practice LLC');
-      expect(result.identifier).toEqual([
-        { system: NPI_SYSTEM, value: '3564119220' },
-        { system: EIN_SYSTEM, value: '123456789' },
-        { system: MEDPLUM_PROVIDER_IDENTIFIER_SYSTEM, value: BILLING_ORGANIZATION_IDENTIFIER_VALUE },
-      ]);
-      expect(result.type).toEqual([
-        {
-          coding: [
-            { system: ORGANIZATION_TYPE_SYSTEM, code: PROVIDER_ORGANIZATION_TYPE, display: 'Healthcare Provider' },
-          ],
-        },
-      ]);
-      expect(result.telecom).toEqual([{ system: 'phone', value: '6175550142' }]);
-      expect(result.address?.[0]?.city).toBe('Boston');
-    });
-
-    test('does not duplicate the prov type or clobber unrelated identifiers and types', () => {
-      const existing: Organization = {
-        resourceType: 'Organization',
-        id: 'org-1',
-        name: 'Old Name',
-        type: [
-          { coding: [{ system: ORGANIZATION_TYPE_SYSTEM, code: PROVIDER_ORGANIZATION_TYPE }] },
-          { coding: [{ system: 'https://example.com/custom-type', code: 'clinic' }] },
-        ],
-        identifier: [
-          { system: 'https://example.com/legacy-id', value: 'legacy' },
-          { system: NPI_SYSTEM, value: '1234567893' },
-        ],
-      };
-
-      const result = buildUpdatedOrganization(existing, {
-        name: 'New Name',
-        npi: '3564119220',
-        ein: '123456789',
-        phone: '',
-      });
-
-      expect(result.id).toBe('org-1');
-      expect(result.type).toHaveLength(2);
-      expect(result.identifier).toEqual([
-        { system: 'https://example.com/legacy-id', value: 'legacy' },
-        { system: NPI_SYSTEM, value: '3564119220' },
-        { system: EIN_SYSTEM, value: '123456789' },
-        { system: MEDPLUM_PROVIDER_IDENTIFIER_SYSTEM, value: BILLING_ORGANIZATION_IDENTIFIER_VALUE },
-      ]);
-      expect(result.telecom).toBeUndefined();
-    });
-
-    test('adds the prov type to an organization that has other types only', () => {
-      const existing: Organization = {
-        resourceType: 'Organization',
-        type: [{ coding: [{ system: 'https://example.com/custom-type', code: 'clinic' }] }],
-      };
-      const result = buildUpdatedOrganization(existing, { name: 'X', npi: '3564119220', ein: '123456789', phone: '' });
-      expect(result.type).toHaveLength(2);
-      expect(
-        result.type?.some((t) =>
-          t.coding?.some((c) => c.system === ORGANIZATION_TYPE_SYSTEM && c.code === PROVIDER_ORGANIZATION_TYPE)
-        )
-      ).toBe(true);
-    });
-
-    test('stamps the provider marker identifier exactly once across repeated saves', () => {
-      const first = buildUpdatedOrganization(
-        { resourceType: 'Organization' },
-        { name: 'X', npi: '3564119220', ein: '123456789', phone: '' }
-      );
-      const second = buildUpdatedOrganization(first, { name: 'X', npi: '3564119220', ein: '123456789', phone: '' });
-      const markers = second.identifier?.filter((id) => id.system === MEDPLUM_PROVIDER_IDENTIFIER_SYSTEM);
-      expect(markers).toEqual([
-        { system: MEDPLUM_PROVIDER_IDENTIFIER_SYSTEM, value: BILLING_ORGANIZATION_IDENTIFIER_VALUE },
-      ]);
+      const identifiers = [{ system: CHC_PAYER_ID_SYSTEM, value: '60054' }];
+      expect(upsertIdentifier(identifiers, CHC_PAYER_ID_SYSTEM, '  ')).toBeUndefined();
     });
   });
